+++ argopy/stores/float/argo_float.py
@@ -27,9 +27,9 @@ else:
 
 
 class ArgoFloat(FloatStore):
-    """Argo GDAC float store
+    """Argo float store
 
-    This store makes it easy to load/read/visualize data for a given float from any GDAC location and netcdf files
+    This store makes it easy to load/read/visualize data for a given float from any GDAC location and netcdf files.
 
     Examples
     --------
@@ -44,10 +44,11 @@ class ArgoFloat(FloatStore):
         af = ArgoFloat(WMO, host='s3')     # Shortcut for s3://argo-gdac-sandbox/pub
 
     .. code-block:: python
-        :caption: Load/read GDAC netcdf files as a :class:`xarray.Dataset` or :class:`netCDF4.Dataset`
+        :caption: Load/read GDAC netcdf files
 
-        af.ls_dataset() # Return a dictionary with all available datasets for this float
-        ds = af.open_dataset('prof') # Use keys from the available datasets dictionary
+        af.ls_datasets() # Return a dictionary with all available datasets for this float
+
+        ds = af.open_dataset('prof') # Use keys from .ls_datasets()
         ds = af.open_dataset('meta')
         ds = af.open_dataset('tech')
         ds = af.open_dataset('Rtraj')
@@ -55,20 +56,46 @@ class ArgoFloat(FloatStore):
 
         ds = af.open_dataset('Sprof', netCDF4=True)  # Return a netCDF4 Dataset instead of an xarray
 
+        af._ls() # Return the full list of dataset files
+
+
+    .. code-block:: python
+        :caption: Load/read GDAC netcdf mono-cycle profile files
+
+        af.ls_profiles() # Return a dictionary with all available mono-cycle profile files (everything under the 'profiles' sub-folder)
+
+        # To load one single file, use keys from af.ls_profiles():
+        ds = af.open_profile(12) # cycle number 12, core file
+        ds = af.open_profile('1D') # cycle number 1, descending core file
+        ds = af.open_profile('B15') # cycle number 15, BGC file
+        ds = af.open_profile('B1D') # cycle number 1, descending BGC file
+        ds = af.open_profile('S28') # cycle number 28, BGC synthetic file
+
+        # To load one or more files, provide cycle number(s) and other attributes:
+        ds_list = af.open_profiles([1,2,3])
+        ds_list = af.open_profiles([1,2,3], direction='D')
+        ds_list = af.open_profiles([1,2,3], dataset='B') # Return 'BGC' B files
+        ds_list = af.open_profiles([1,2,3], dataset='B', direction='D') # Return 'BGC' B files, descending
+        ds_list = af.open_profiles([1,2,3], dataset='S') # Return 'BGC' Synthetic files
+
+        # If you don't specify cycle numbers, all cycles are loaded:
+        ds_list = af.open_profiles(direction='D') # Return *all* core descending files
+
+        af.profiles_to_dataframe()  # Pandas DataFrame describing all available profile files
+
+
     .. code-block:: python
         :caption: Other attributes and methods
 
-        af.N_CYCLES  # Number of cycles (estimated)
+        af.CYCLE_NUMBERS  # List of unique cycle numbers (as given by file names under 'profiles' GDAC folder)
+        af.N_CYCLES  # Number of cycles
         af.path  # root path for all float datasets
         af.dac   # name of the DAC this float belongs to
         af.metadata  # a dictionary with all available metadata for this file (from netcdf or fleetmonitoring API)
-        af.ls()  # list af.path folder content
 
-    .. code-block:: python
-        :caption: Working with float profiles
+        af._ls()  # list af.path folder content
+        af._lsp() # list af.path+'profiles' folder content
 
-        af.lsprofiles() # list float "profiles" folder content
-        af.describe_profiles()  # Pandas DataFrame describing all available float profile files
 
     .. code-block:: python
         :caption: Quick plotting methods